requirements.txt
requests
pandas
python-dotenv
tqdm
google-generativeai